load-real-sf-contacts/hello_world/requirements.txt
requests
pandas
simple_salesforce
snowflake-connector-python